update deleteUser now cant delete Admin users

+++ dashboard/src/Redux/Actions/UserActions.jsx
@@ -103,6 +103,38 @@ export const listUser = () => async (dispatch, getState) => {
   }
 };
 
+// export const deleteUser = (id) => async (dispatch, getState) => {
+//   try {
+//     dispatch({ type: USER_DELETE_REQUEST });
+
+//     const {
+//       userLogin: { userInfo },
+//     } = getState();
+
+//     const config = {
+//       headers: {
+//         Authorization: `Bearer ${userInfo.token}`,
+//       },
+//     };
+
+//     await axios.delete(`${URL}/api/users/${id}`, config);
+
+//     dispatch({ type: USER_DELETE_SUCCESS });
+//   } catch (error) {
+//     const message =
+//       error.response && error.response.data.message
+//         ? error.response.data.message
+//         : error.message;
+//     if (message === "Not authorized, token failed") {
+//       dispatch(logout());
+//     }
+//     dispatch({
+//       type: USER_DELETE_FAIL,
+//       payload: message,
+//     });
+//   }
+// };
+
 export const deleteUser = (id) => async (dispatch, getState) => {
   try {
     dispatch({ type: USER_DELETE_REQUEST });
@@ -116,6 +148,17 @@ export const deleteUser = (id) => async (dispatch, getState) => {
         Authorization: `Bearer ${userInfo.token}`,
       },
     };
+
+    // Retrieve the user to be deleted
+    const { data: userToDelete } = await axios.get(
+      `${URL}/api/users/${id}`,
+      config
+    );
+
+    // Check if the user to be deleted is an admin
+    if (userToDelete.isAdmin) {
+      throw new Error("Not authorized to delete admin users.");
+    }
 
     await axios.delete(`${URL}/api/users/${id}`, config);
